pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "archive-to-images"
version = "0.1.0"
description = "Awesome `archive-to-images` is a Python cli/package created with https://github.com/TezRomacH/python-package-template"
readme = "README.md"
authors = ["Giovanni Pecoraro <giovanni1.pecoraro@protonmail.com>"]
license = "MIT"
repository = "https://github.com/Peco602/archive-to-images"
homepage = "https://github.com/Peco602/archive-to-images"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"archive-to-images" = "archive_to_images.__main__:app"

[tool.poetry.dependencies]
python = "^3.8"
importlib_metadata = {version = "^4.5.0", python = "<3.8"}
typer = {extras = ["all"], version = "^0.6.1"}
rich = "^12.6.0"
Pillow = "9.3.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.4"
darglint = "^1.8.1"
black = {version = "^22.10.0", allow-prereleases = true}
isort = {extras = ["colors"], version = "^5.10.1"}
mypy = "^0.982"
pre-commit = "^2.20.0"
pydocstyle = "^6.1.1"
pylint = "^2.15.5"
pytest = "^7.2.0"
pyupgrade = "^3.2.0"
safety = "^2.3.1"
coverage = "^6.5.0"
coverage-badge = "^1.1.0"
pytest-html = "^3.2.0"
pytest-cov = "^4.0.0"
pyfakefs = "5.0.0"

[tool.black]
# https://github.com/psf/black
target-version = ["py37"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 37
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.8
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "archive-to-images"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true
